requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
openai==0.27.6
packaging==23.1
Pillow==9.5.0
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.4
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-kernel==0.2.6.dev0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
